pyproject.toml
[project]
name = "schellenberg_usb"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "pyschellenberg>=0.1.0",
]

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" }
]

test = [
    "croniter~=6.0.0",
    "watchdog~=6.0.0",
    "mock-open~=1.4.0",
    "polyfactory~=3.1.0",
    "PyYAML~=6.0.2",
    # Need specific versions
    "pytest-homeassistant-custom-component==0.13.299",
    # Versions handled by pytest-homeassistant-custom-component
    "freezegun",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

lint = [
    "mypy~=1.18.2",
    "pre-commit~=4.5.0",
    "ruff~=0.14.1",
    "types-PyYAML~=6.0.12.20250915",
]

[tool.uv.sources]
pyschellenberg = { path = "pyschellenberg", editable = true }


[tool.coverage.run]
source = ["custom_components"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplemented()",
    "if __name__ == '__main__':",
    "main()",
]
show_missing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git"]
addopts = """
-n4
--strict-markers
--cov=custom_components"""
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore:Inheritance class CountingClientSession from ClientSession is discouraged:DeprecationWarning",
]

[tool.mypy]
python_version = "3.13"
pretty = true
follow_imports = "silent"
ignore_missing_imports = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.ruff.lint.pycodestyle]
max-line-length = 80